+++ src/pages/ShoppingCart/ShoppingCart.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { productsPC } from "../../data/ProductsPCData";
 import "./../../static/styles/ShoppingCart.css";
-import { deliveryMethods } from "../../data/DeliveryMethodData";
-import { DeliveryAddressForm } from "../../forms/ShoppingCart/DeliveryAddressForm";
+import { DeliveryAddressForm } from "./components/DeliveryAddressForm";
+import { DeliveryMethod } from "./components/DeliveryMethod";
 
 export const ShoppingCart = () => {
     const selectedIds = [
@@ -99,47 +99,8 @@ export const ShoppingCart = () => {
                             <div>Total: $ {totalPrice}</div>
                         </div>
                     </div>
-                    <div className="card mb-2">
-                        <h5 className="card-header">Delivery method</h5>
-                        <div className="card-body">
-                            {deliveryMethods.map((data, index) => (
-                                <div key={index} className="mb-2">
-                                    <div className="form-check d-flex align-items-center">
-                                        <input
-                                            className="form-check-input"
-                                            type="radio"
-                                            name="deliveryMethod"
-                                            id={`deliveryMethod${index}`}
-                                        />
-                                        <label
-                                            htmlFor={`deliveryMethod${index}`}
-                                            className="form-check-label ms-2 w-100"
-                                        >
-                                            <div className="card">
-                                                <div className="card-body">
-                                                    <h5 className="card-title">
-                                                        {data.name}
-                                                    </h5>
-                                                    <h6 className="card-subtitle mb-2 text-muted">
-                                                        {data.info}
-                                                    </h6>
-                                                    <h6 className="card-subtitle mb-2 text-muted">
-                                                        {data.price}
-                                                    </h6>
-                                                </div>
-                                            </div>
-                                        </label>
-                                    </div>
-                                </div>
-                            ))}
-                        </div>
-                    </div>
-                    <div className="card mb-2">
-                        <h5 className="card-header">Delivery address</h5>
-                        <div className="card-body">
-                            <DeliveryAddressForm />
-                        </div>
-                    </div>
+                    <DeliveryMethod />
+                    <DeliveryAddressForm />
                 </div>
                 <div className="d-flex justify-content-center mb-3">
                     <a
